chore: rm `wasm_bindgen_test` from core-crypto
Core crypto doesn't know or care about the environment it's compiled for;
it's intentionally independent of its target. `crypto-ffi` has to know
about things like WASM, and run its test suite on that target. These
tests intrinsically test core-crypto.

There's no point in doing the relatively expensive process of spinning
up a wasm environment and compiling core-crypto for wasm for testing;
this just validates against miscompilation, which is not a reasonable
thing to guard against, nor would it be solvable on our part if we did
find it.

So we get rid of it all.