refactor(core-crypto): simplify mls session constructor/config
There were two different constructors with a complicated config
struct, whose fields were sometimes required but ignored. This
config struct claimed to require validation but could be manually
constructed, bypassing the validation.

Now there is one constructor and one config struct with a builder.
This config struct is validated in a typesafe way which cannot be
bypassed.