refactor(interop/web): define JS functions in real TS files
This isn't a perfect situation, but it's a massive improvement
over the status quo. Typescript is smart enough to point out when
we're using one of our interfaces wrong even given the pure-js
implementation required by the fact that we attempt no transpilation.

One area we do need to improve: right now, the implementation
re-parses the source file every time the function is called.
Parsing is fast, but not _that_ fast. So we should fix that.