build(keystore): do not attempt to build `smol` when testing in wasm
`smol` doesn't work on wasm32-unknown-unknown, for reasons including
but not necessarily limited to a dependency on the `errno` crate,
which is meaningless without an OS to generate error numbers.

This required some minor changes to the keystore, as its test setup
requires an executor on which one can block. Luckily, we already
had an alternative among our dependencies.

We also had to manually adjust the test attribute according to the
target family within the `all_storage_types` template. Once that
was done, everything basically worked.