build: replace `tokio` with `smol` in direct deps, except interop
It turns out that while `smol` is fine for abstract async execution,
the rust web development community has as a whole chosen `tokio` for
servers. This is relevant because interop runs a tiny little server
for the wasm stuff.

I could not find _any_ approach higher-level than a `TcpStream`
which did not require Tokio, had been updated in the last 3 years,
and had a non-trivial userbase. There are various personal projects,
and there are old projects that generate supply-chain errors, and
then there's Tokio.

As long as interop needs to provide a web server, I think we're
stuck with Tokio.