ci: make release process atomic [WPB-16043]
We're now waiting for the artifacts of all platforms to be successfully
produced before publishing for any platform. Those artifacts will first
be uploaded to a GitHub release. Then we'll try publishing those
artifacts on their respective platforms in parallel. If any of that
fails, we can simply retry (manually) with the artifacts published on
GitHub.